requirements.txt
astroid==2.3.2
cachetools==3.1.1
certifi==2019.9.11
dal-admin-filters==0.3.3
Django==2.2.7
django-admin-vali==0.2.2
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-debug-toolbar==2.1
django-imagekit==4.0.2
django-leaflet==0.25.0
django-map-widgets==0.2.2
django-material==1.5.8
django-material-admin==1.5.0
django-smart-selects==1.5.4
django-taggit==1.1.0
GDAL==2.3.3
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
httplib2==0.14.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.3
pilkit==2.0
Pillow==6.2.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pylint==2.4.3
pytz==2019.3
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
wrapt==1.11.2
